requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-filter==2.2.0
django-rest-knox==4.1.0
django-simple-history==2.8.0
djangorestframework==3.11.0
psycopg2-binary==2.8.4
pycparser==2.19
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
